fix(performance): Move useEffect Todo Context
